requirements.txt
pandas
gym
pyglet
tensorflow
keras
keras-rl2
stable-baselines3[extra]
python-binance
mplfinance
ta